Move card models and get info request factory and parser to CorePayments
