+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardClient.kt
@@ -7,13 +7,13 @@ import com.braintreepayments.api.BrowserSwitchOptions
 import com.braintreepayments.api.BrowserSwitchResult
 import com.braintreepayments.api.BrowserSwitchStatus
 import com.paypal.android.cardpayments.api.CardAPI
-import com.paypal.android.cardpayments.api.GetOrderRequest
-import com.paypal.android.cardpayments.model.CardResult
+import com.paypal.android.corepayments.models.CardResult
 import com.paypal.android.corepayments.API
 import com.paypal.android.corepayments.APIClientError
 import com.paypal.android.corepayments.CoreConfig
 import com.paypal.android.corepayments.CoreCoroutineExceptionHandler
 import com.paypal.android.corepayments.PayPalSDKError
+import com.paypal.android.corepayments.api.models.GetOrderRequest
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
