+++ CorePayments/src/main/java/com/paypal/android/corepayments/api/models/GetOrderRequest.kt
@@ -1,0 +1,3 @@
+package com.paypal.android.corepayments.api.models
+
+data class GetOrderRequest(val orderId: String)
