+++ CorePayments/src/main/java/com/paypal/android/corepayments/models/AuthenticationResult.kt
@@ -1,6 +1,5 @@
-package com.paypal.android.cardpayments.model
+package com.paypal.android.corepayments.models
 
-import com.paypal.android.cardpayments.threedsecure.ThreeDSecureResult
 import com.paypal.android.corepayments.PaymentsJSON
 import org.json.JSONObject
