+++ CorePayments/src/main/java/com/paypal/android/corepayments/models/PaymentSource.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.cardpayments.model
+package com.paypal.android.corepayments.models
 
 import com.paypal.android.corepayments.PaymentsJSON
 import org.json.JSONObject
@@ -17,7 +17,7 @@ data class PaymentSource(
         const val KEY_AUTHENTICATION_RESULT = "authentication_result"
     }
 
-    internal constructor(json: PaymentsJSON) : this(
+    constructor(json: PaymentsJSON) : this(
         json.getString(KEY_LAST_DIGITS),
         json.getString(KEY_BRAND),
         json.optString(KEY_TYPE),
